Removes the check for full readiness of 3 replicas
In the ci environment the amount of time between 3 pods being created
and all being ready seems highly variable.  Also since we currently lack
the status information to determine if a deployment is not progressing,
it's best just to remove this check.  The latter check asserting full
readiness of 2 replicas will suffice.

Closes #20888